requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.1.0
python-decouple==3.8
djangorestframework-simplejwt==5.3.1
whitenoise==6.6.0
dj-database-url==2.2.0
django-cloudinary-storage==0.3.0
cloudinary==1.40.0
python-dotenv==1.0.0
psycopg2-binary
gunicorn==21.2.0

